Fixed some parsing and tokenization errors.  Exceptions are no longer thrown to signify tokenization, parsing, or evaluation errors; NSErrors are used instead.
Tokenization is no longer locale aware.  Numbers must use "." as the decimal separator and "e" or "E" as the exponent symbol.  Numbers using "," will not be recognized.  This stems from attempting to parse "min(1,2,3)".  In locales where "," is the decimal separator, this is ambiguous.  It could be any one of:
min(1, 2, 3)
min(1.2, 3)
min(1, 2.3)

The inability to distinguish between these requires using "." as the decimal separator.