Fix unit tests assuming right precedence of pow.
The actual implementation determines the precedence of the power
operator during runtime, but the tests use a hard coded value. As a
result, the assertion that "2**3**2 == 512" on runtimes which use the
correct, left-precedence evaluation order (such as my OS X 10.9
MacBook). This commit adds the same runtime check to the tests in order
to prevent false negatives in the tests.