ignore unary plus when expressionizing
The unary plus operator ("+3") is a worthless operator and has no effect on the final evaluation of the expression. However, it's useful to keep it around for as long as possible, in the interest of providing more information to clients of the various token stages. Thus, we should only strip out the unary plus operator when it comes time to actually turn the tokens into the final Expression.

Fixes issue #104