Fixes issue #145
Unary operators tend to have the same precedence, and are just evaluated in inner-to-outer order. However, if you happen to have a group where there are multiple unary operators, then it would extract the first one from the group, but potentially use it at another index. Instead now, when we have multiple operators with the same precedence, we'll use the first operator's associativity to determine which index to use, but then use that index to re-extract the operator from the array of tokens, so we don't accidentally substitute an operator in to the wrong spot.